+++ public/app/features/explore/QueryRow.tsx
@@ -23,7 +23,7 @@ import {
 import { selectors } from '@grafana/e2e-selectors';
 
 import { ExploreItemState, ExploreId } from 'app/types/explore';
-import { highlightLogsExpressionAction } from './state/exploreItem';
+import { highlightLogsExpressionAction } from './state/explorePane';
 import { ErrorContainer } from './ErrorContainer';
 import { changeQuery, modifyQueries, removeQueryRowAction, runQueries } from './state/query';
